float update

+++ fairseq/data/indexed_dataset.py
@@ -90,7 +90,7 @@ dtypes = {
     3: np.int16,
     4: np.int32,
     5: np.int64,
-    6: np.float64,
+    6: np.float32,
     7: np.double,
     8: np.uint16
 }
@@ -298,7 +298,7 @@ class IndexedDatasetBuilder(object):
         np.int16: 2,
         np.int32: 4,
         np.int64: 8,
-        np.float: 4,
+        np.float32: 4,
         np.double: 8
     }
